+++ src/bos/operators/utils/clients/bos/sessions_status.py
@@ -1,7 +1,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2021-2022, 2024 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2021-2025 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -21,14 +21,16 @@
 # ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 # OTHER DEALINGS IN THE SOFTWARE.
 #
+from functools import partial
 import json
 import logging
 
 from .base import BASE_ENDPOINT, log_call_errors
 from bos.operators.utils import requests_retry_session
+from .base import check_bos_response as _check_bos_response
 
 LOGGER = logging.getLogger('bos.operators.utils.clients.bos.sessions_status')
-
+check_bos_response = partial(_check_bos_response, logger=LOGGER)
 
 class SessionStatusEndpoint(object):
     ENDPOINT = 'sessions'
@@ -43,7 +45,7 @@ class SessionStatusEndpoint(object):
         session = requests_retry_session()
         LOGGER.debug("GET %s", url)
         response = session.get(url)
-        response.raise_for_status()
+        check_bos_response(response)
         item = json.loads(response.text)
         return item
 
@@ -57,6 +59,6 @@ class SessionStatusEndpoint(object):
         url = self.base_url + '/' + session_id + '/status'
         LOGGER.debug("POST %s", url)
         response = session.post(url)
-        response.raise_for_status()
+        check_bos_response(response)
         items = json.loads(response.text)
         return items